CollisionManager 추가 충돌을 처리하기 위한 매니저

+++ WinApi2D/struct.h
@@ -67,7 +67,7 @@ struct fVec2
 
 	fVec2& normalize()
 	{
-		float length = (float)sqrt(x * x + y * y);
+		float length = (float)sqrt((double)x * x + (double)y * y);
 
 		assert(length != 0.f);
